+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/impl/DataFetchersDelegateNodeImpl.java
@@ -5,13 +5,14 @@ import java.util.UUID;
 
 import org.allGraphQLCases.server.Node;
 import org.allGraphQLCases.server.util.DataFetchersDelegateNode;
+import org.dataloader.BatchLoaderEnvironment;
 import org.springframework.stereotype.Component;
 
 @Component
 public class DataFetchersDelegateNodeImpl implements DataFetchersDelegateNode {
 
 	@Override
-	public List<Node> batchLoader(List<UUID> keys) {
+	public List<Node> batchLoader(List<UUID> keys, BatchLoaderEnvironment environment) {
 		throw new RuntimeException("Not implemented yet");
 	}